refactor(webapp): pay down boy-scout debt in secret-backends.ts
Move secrets bridge client to the shell transport layer (mirroring
float-topology) so secret-backends no longer imports up-stack into core.
Replace open Record<string, unknown> message bags with a named
SecretsControlMessage union and ratchet both debt baselines.
